Move device automator code to com.braintreepayments package.
