+++ DeviceAutomator/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
-<manifest package="com.lukekorth.deviceautomator">
+<manifest package="com.braintreepayments.deviceautomator">
 </manifest>
