+++ DeviceAutomator/src/test/java/com/braintreepayments/UiObjectMatcherTest.java
@@ -1,4 +1,4 @@
-package com.lukekorth.deviceautomator;
+package com.braintreepayments;
 
 import androidx.test.uiautomator.UiDevice;
 import androidx.test.uiautomator.UiObject;
@@ -6,6 +6,8 @@ import androidx.test.uiautomator.UiSelector;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import com.braintreepayments.UiObjectMatcher;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
