+++ DeviceAutomator/src/test/resources/robolectric.properties
@@ -1,2 +1,2 @@
 sdk=21
-constants=com.lukekorth.deviceautomator.BuildConfig
+constants=com.braintreepayments.deviceautomator.BuildConfig
