+++ README.md
@@ -18,7 +18,7 @@ Add the dependency in your `build.gradle` file:
 
 ```groovy
 dependencies {
-  androidTestCompile 'com.lukekorth:device-automator:1.0.0'
+  androidTestCompile 'com.braintreepayments:device-automator:1.0.0'
 }
 ```
 
@@ -26,7 +26,7 @@ To use the latest build from the `master` branch use:
 
 ```groovy
 dependencies {
-  androidTestCompile 'com.lukekorth:device-automator:1.0.0-SNAPSHOT'
+  androidTestCompile 'com.braintreepayments:device-automator:1.0.0-SNAPSHOT'
 }
 ```
